BookDetails - delete dialogue

+++ app/src/main/java/com/example/semestdrahosvamz/ui/screens/details/BookDetailsScreen.kt
@@ -6,19 +6,17 @@ import android.net.Uri
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
-import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.width
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Add
 import androidx.compose.material.icons.filled.ArrowBack
 import androidx.compose.material.icons.filled.Delete
+import androidx.compose.material3.AlertDialog
 import androidx.compose.material3.ExperimentalMaterial3Api
 import androidx.compose.material3.FloatingActionButton
 import androidx.compose.material3.Icon
@@ -28,6 +26,7 @@ import androidx.compose.material3.RadioButton
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
 import androidx.compose.material3.TopAppBar
 import androidx.compose.material3.TopAppBarDefaults
 import androidx.compose.runtime.Composable
@@ -37,16 +36,18 @@ import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.asImageBitmap
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.viewmodel.compose.viewModel
 import com.example.semestdrahosvamz.Data.Book
+import com.example.semestdrahosvamz.R
 import com.example.semestdrahosvamz.ui.ViewModelProvider
-import com.example.semestdrahosvamz.ui.screens.bookEntry.BookEntryViewModel
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 
@@ -102,7 +103,7 @@ fun BookBaseInfo(book: Book, innerPadding: PaddingValues, onStatusChange : (Int)
 
 @Composable
 fun ReadingStatusOptions(selectedOption: Int, onOptionSelected: (Int) -> Unit) {
-    val options = listOf("Finished", "Reading", "Planned")
+    val options = listOf(stringResource(R.string.readingStatusFinished), stringResource(R.string.readingStatusReading), stringResource(R.string.readingStatusPlanned))
     options.forEachIndexed { index, text ->
         Row(
             verticalAlignment = Alignment.CenterVertically,
@@ -121,8 +122,21 @@ fun ReadingStatusOptions(selectedOption: Int, onOptionSelected: (Int) -> Unit) {
     }
 }
 
-fun OnDeletion(navigateBack: () -> Unit) {
-    navigateBack
+@Composable
+fun DeleteDialogue(onConfirm : () -> Unit, onCancel : () -> Unit) {
+    AlertDialog(onDismissRequest = { /* Do nothing */ },
+        title = { Text(stringResource(R.string.delDiaTitle)) },
+        text = { Text(stringResource(R.string.delDiaText)) },
+        dismissButton = {
+            TextButton(onClick = onCancel) {
+                Text(stringResource(R.string.no))
+            }
+        },
+        confirmButton = {
+            TextButton(onClick = onConfirm) {
+                Text(stringResource(R.string.yes))
+            }
+        })
 }
 
 @OptIn(ExperimentalMaterial3Api::class)
@@ -130,15 +144,17 @@ fun OnDeletion(navigateBack: () -> Unit) {
 fun BookDetailsScreen(navigateBack: () -> Unit, viewModel: BookDetailsViewModel = viewModel(factory = ViewModelProvider.Factory)) {
     val coroutineScope = rememberCoroutineScope()
     val uiState = viewModel.uiState.collectAsState()
+    var deleteDialogue by rememberSaveable { mutableStateOf(false) }
+
     Scaffold(
         topBar = {
             TopAppBar(
                 title = {
-                    Text(text = "Place holder add book")
+                    Text(text = uiState.value.book.title)
                 },
                 navigationIcon = {
                     IconButton(onClick = navigateBack) {
-                        Icon(Icons.Default.ArrowBack, contentDescription = "Back")
+                        Icon(Icons.Default.ArrowBack, contentDescription = "")
                     }
                 },
                 colors = TopAppBarDefaults.topAppBarColors(
@@ -149,16 +165,25 @@ fun BookDetailsScreen(navigateBack: () -> Unit, viewModel: BookDetailsViewModel 
         },
         floatingActionButton = {
             FloatingActionButton(
-                onClick = {
-                    coroutineScope.launch {
-                        viewModel.deleteBook()
-                        navigateBack()
-                    }},
+                onClick = {deleteDialogue = true}
             ) {
                 Icon(Icons.Filled.Delete, "")
             }
         },
 
-    ) { innerPadding -> BookBaseInfo(book = uiState.value.book, innerPadding = innerPadding, viewModel::updateReadingStatus)
+    ) { innerPadding ->
+        BookBaseInfo(book = uiState.value.book, innerPadding = innerPadding, viewModel::updateReadingStatus)
+        if (deleteDialogue) {
+            DeleteDialogue(
+                onConfirm = {
+                    coroutineScope.launch {
+                        deleteDialogue = false
+                        viewModel.deleteBook()
+                        navigateBack()
+                    }
+                },
+                onCancel = {deleteDialogue = false}
+                )
+        }
     }
 }
